# Optimised Bytes Array: Conditional buffer based on Vectorisation