Fixes https://github.com/datumbox/VernamVeil/issues/41

After implementing https://github.com/datumbox/VernamVeil/pull/42, I realised I could potentially get the same effect without removing the Scalar API. We could just have the type of the buffer depend on the vectorisation flag of the FX:
- In non-vectorised mode, bytearrays will be used.
- In vectorised mode, numpy arrays will be used.

The benchmarks show we can get the same speed without losing the scalar API.

Previous baseline from #40:
```
The 'encode' step took 4.313 seconds.
The 'decode' step took 4.183 seconds.
```

Using https://github.com/datumbox/VernamVeil/pull/42 (782fa1d) and removing the scalar API:
```
The 'encode' step took 3.307 seconds.
The 'decode' step took 3.356 seconds.
```

Maintaining the scalar API but just adopting conditional buffers (39cb8b8):
```
The 'encode' step took 3.321 seconds.
The 'decode' step took 3.352 seconds.
```

Pretty much the same!

Though simplifying the code and removing the scalar API is not off the table, from Memory management perspective, we can decouple these decision. We will merge this PR instead of #42 as we literally incur no penalty by keeping the scalar API and revise it on the future.